Auth (#2)
* prefs
* write user to db on login
* env
* CollectionTypeIndicators
* decent auth

+++ app/src/main/java/net/treelzebub/studiopeer/model/entities/Track.kt
@@ -31,7 +31,7 @@ interface Track : DatabaseObject {
     /**
      * Length in milliseconds.
      */
-    var length: Int
+    var length: Long
 
     /**
      * File size in bytes.
